requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
blueprint==3.4.2
certifi==2018.11.29
chardet==3.0.4
Click==7.0
colorama==0.4.0
coverage==4.5.2
coveralls==1.5.1
docopt==0.6.2
Flask==1.0.2
Flask-JWT==0.3.2
Flask-JWT-Extended==3.14.0
gunicorn==19.9.0
idna==2.8
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
more-itertools==4.3.0
nose==1.3.7
pluggy==0.8.0
psycopg2==2.7.6.1
py==1.7.0
PyJWT==1.4.2
pylint==2.1.1
pytest==4.0.1
pytest-cov==2.6.1
requests==2.21.0
six==1.11.0
urllib3==1.24.1
Werkzeug==0.14.1
wrapt==1.10.11
